Initial configuration for FLSUN Q5
This is based on the configuration and pins files from alexeyzel's repo. This enables the new color touch UI that was just added to Marlin bugfix 2.0.x.

This config includes delta and PID tuning for my personal Q5, as well as a temporary workaround to prevent initial hot end temp overshoot (see notes in Configuration.h).