Added additional serial port to connect TFT screen
Previously the TFT screen was connected to the same serial port that is shared by the USB port. This caused issues when connecting Octoprint at the same time as the messages sent from Marlin would be received by both the screen and Octoprint.

I enabled serial port 2 (pins D16 and D17 no the MKS Gen L board) and rewired the screen to connect to those pins instead of the serial port 0 pins. Now Octoprint (or a computer) can connect to the USB port and control the printer without the screen interfering.

This should also help with Marlin flashing, though I haven't confirmed that yet.